qtmod_test: raise a fatal error on dirty build directory
If the build directory has not been properly cleaned and still contains
old copies of some repositories, "git clone" will give a fatal error and
the build will abort.

Explicitly check it ourselves, allowing us to get more detail in the
error message. This also allows parse_build_log to detect and retry on
this error.

This error is most commonly observed on Windows machines when a process
is unexpectedly holding a handle to some file under the build directory.

Change-Id: I734c3001ddf828b4af74ce051646873d538068e5
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>