Differentiate between failures and errors when aborting a test script
Previously, fatal_error() was the only way to abort a script with a
message to be understood by parse_build_log.pl; however, it makes sense
to differentiate between "failure" (we probably found a genuine problem
in the software under test) and "error" (something unexpected happened
which prevents us from continuing, probably unrelated to the software
under test).

The main reason to categorize problems as either failures or errors is
that it often makes sense to retry and/or notify some responsible CI
person on errors, but rarely makes sense to do that on failures.

Change-Id: I3b4d71f0431ed06b864b6f576658ada097f13e33
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>