parse_build_log: retry on testconfig or git clone errors
If "determining test script configuration" fails, it usually means that
someone has made a mistake in qtqa/testconfig.

If "setting up git repositories" fails, it usually means that some
temporary outage has occurred in the network or the git server.

In both cases, it's unrelated to the code under test and it makes
sense to retry again a bit later, so do that.

Change-Id: I517a74cbddce8cf5a51d91a209b9d45416158c4c
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>