parse_build_log: truncate lines only once, as they are read
We were reading the lines at the beginning of the script, then
truncating them twice in two different parts of the script.

It's faster to instead only truncate the lines in one place, as they are
read, and not using an additional function call with pass-by-reference
on each line.

Reduces average runtime by ~3% (for me).

Change-Id: I5112db259e7cd60953f8fb10f2de0f35a4afcbbc
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>