qt-jenkins-integrator: minor improvements to email configuration
Let the Subject prefix be configurable, and default to blank; it doesn't
make much sense to prepend [Qt CI] if the mailing list itself already
does something similar.

Let Reply-To be configurable. The recommended setup is to set it equal
to the mailing list address so that replying to a CI report replies to
the list by default.

Change-Id: I8c11cdf103a53bc2944e1379c882280cce86afff
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>