qt-jenkins-integrator: upload stash in JSON format next to build logs
When a build is completed, upload (a subset of) the integrator's stash
for the project, into a state.json.gz file.

This contains valuable information such as the build ref in gerrit, the
changes being tested, and the overall build result. Previously, this
data hasn't been exported on testresults.qt-project.org in any
machine-readable manner.

Change-Id: I8fac8ac2dc7163806568fc719f658227dc4b507d
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>