Add admin interface to qt-jenkins-integrator.pl
Allow admin commands to be sent to qt-jenkins-integrator.pl via TCP.
A TCP port is configured with AdminTcpPort which listens for
incoming connections and admin commands as JSON objects.
The commands are authenticated with JenkinsToken

Supports remove-state and reset-state commands that can be used
to handle error conditions without stopping integrator script.

Task-number: QTQAINFRA-577

Change-Id: I0e8c76b7d467c511c603387d297812ec8e0d3da2
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Rohan McGovern <rohan@mcgovern.id.au>
Reviewed-by: Simo Fält <simo.falt@digia.com>