Fix qtqa autotest for mingw build.
Skip autotest in 41-print_calling_makefile.t if MinGW is used.
This script is a workaround for nmake's lack of any equivalent
to GNU make's $(MAKEFILE_LIST), thus there is no need to test
it with MinGW builds.

Pass correct make command for testplanner.pl and testscheduler.t,
in respective test cases. Without '--make' argument, the scripts
default to `nmake' on Windows and `make' everywhere else.

Change preprocessor macro used in pass.cpp, so that file compiles
for both Visual Studio and MinGW. According to MS docs
http://msdn.microsoft.com/en-us/library/b0084kay(v=vs.80).aspx
is always defined for Windows applications (both 32 and 64-bit).

Change-Id: I3a63a8c44b4bb0d0c066d8dc21a6559d54cfdb98
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>