qt-jenkins-deploy: added variable for boot script.
Added variable boot_script for entering Groovy script that does
the machine reboot after build/test. The script is txt file that
resides on same location as the .ini file that includes it. Boot
script is used in template file with [% boot_script %].
This removes the need of copy/pasting the actual boot script to several
templates making it hard to maintain.

Change-Id: I31a4f38122abf77f125d760ae3450dfc78ca8fc9
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>