Update the test zone to use test IP addresses.
Using Class E addresses was an interesting idea, but foiled by
Microsoft Windows trying to be helpful. Empirical evidence shows that
the getaddrinfo() call on Windows filters addresses in the 240.0.0.0/4
and 0.0.0.0/8 ranges. I couldn't be certain, but it seemed to filter
the link-local IPv6 addresses too (fec0::/16).

At least the test networks aren't filtered. For now...

Change-Id: I613d99dff656526921731aa293318264f76383e6
Reviewed-by: Richard J. Moore <rich@kde.org>