Fetch headers according to sync.profile.
This avoids the test checking headers under src/ that won't be
installed. This will be useful for QtWebEngine where most of the
headers are part of the adaptation layer between Qt and Chromium
while the exported symbols are restricted to the subdirectory
specified in sync.profile.

This however stops checking for plugins and other dynamic libraries
that could still benefit from this check to prevent polluting the
namespace for dynamic symbols not part of the public API, but still
with headers not ending with "_p.h".

Change-Id: Id90e14a7a2a11c83c69e723faf2f3bbe2997a4d5
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>