qtmod_test: Fix run_install_check() when cross compilating
run_install_check() assumes that we are doing a Desktop build so it
expects the /include/ directory to end up under $(qt.install.dir)/include
which is not always true, especially when you are cross compiling and using
-sysroot and/or -headerdir which may place the /include/ directory
somewhere else.

Let's check for 'bin' and 'mkspecs' instead, as it's more likely
they end up under $(qt.install.dir) as they are host related artifacts.

Change-Id: I29463fa73916d08880bb5f15fa1b6ca721511c43
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>