Testconfig_enforcer reacts to 403 HTTP returns
The code was done to react to HTTP 404 error messages
so that the script returned an empty string. But if
some other HTTP error message was received, the script
died. To make the script more user frindly, we can
consider the error message 403 the same and return
an empty string here too.

Change-Id: I4940170699c713c96048662ac5fc14661d5dc5a3
Reviewed-by: Heikki Halmet <heikki.halmet@digia.com>
Reviewed-by: Simo Fält <simo.falt@digia.com>