add branch_qt script
this script creates branches in all qt submodules and the supermodule.

additionally, it bumps the version number in the parent branch, and
direct-pushes it. that's ok, because during branching, integration is
(supposed to be) locked down. omitting the review for these trivial
automated changes saves everybody (thusly not) involved some RSI.

Change-Id: I39af889ae5443b002e949a83fa3505b65cbf0a40