Skip license header tests in qtqa
Qtqa module isn't branched and so on it isn't sure if it has proper
header templates in use for each branch. At the moment qt5.git
integration is failing in '5.4' because qtqa is using newer header
templates. We should either branch the qtqa or skip header tests there.
It is easier (&safe) to skip the test.

Change-Id: If573d00094ddfa9515a47960427c37a284379310
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>