Add support for incremental license changes to the autotest
It's currently extremely difficult to update an existing license
text from an old to a new version as the header check will block
these if the license type after QT_BEGIN_LICENSE is the same as
before.

So also check for header.LICENSE_TYPE-OLD, and if that matches
pass the test but print out a warning about using an old,
deprecated version of the license text.

Change-Id: I10d37913a1772887aed7e34e3048325622fd07fc
Reviewed-by: Edward Welbourne <edward.welbourne@theqtcompany.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>