qtestcompare: Add handling of callgrind instruction counts.
To ensure that results make sense, we only compare things that match up:
duration vs duration, or instruction count vs instruction count. If data is
missing, we still won't attempt to compare, too.

Anything that does not make sense will not be compared (but will be printed if
present).

It may be that instruction count should not be stored/printed as a float:
this is a change for another day.

Done-with: Simo Fält <simo.falt@theqtcompany.com>
Change-Id: Ia35e2339d384f971b76166ba2a19c202c5c5fd92
Reviewed-by: Simo Fält <simo.falt@theqtcompany.com>