qtestcompare: Fix percentages and incorrect direction of verdict.
Calculated values were:
(new - old) / new, which gives us a ratio, but to turn that into a percentage we
need to multiply by 100.

Also, if we have a negative percentage, there's a reduction (in time or
instructions) which is a good thing, not a bad thing -- the calculation was the
wrong way around.

Change-Id: I56f165cc22c0f530c226ce7a569e79c7818c1a7f
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>