Added little helper tool to create an initial changelog template for a module
Simply run the binary in the module directory with the correct branch checked
out.  The tool will peek at .qmake.conf to figure out the current version and
run git tag -l to see what the previous version is.

The resulting change log requires manual editing, but it is a start.

Change-Id: I975c0d7a74fee8cab2ae6f400972c5dbc73ff367
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>