Iterate modules in the _data() method, not tst_Bic::initTestCases()
Calling QTest::addColumn() and QTest::newRow() in initTestCases() used
to iterate the resulting value, accessed via QFETCH_GLOBAL(), as a
prefix to all tests. Recently it changed to only doing one of the
values. This is used in some tests in qtbase without problems, usually
in initTestCases_data(), but sometimes in initTestCases(); it's not
clear why it fails here.

In any case, there's no need for this test to not do the iteration
over libraries in the _data() method, so move it there and use these
QTest methods in their usual way, with QFETCH to pick up the datum.

Task-number: QTBUG-61774
Change-Id: I11f7772cf4c83e1133cb2e5cfc7d8230b900892c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>