Add production baseline scripts
Baseline creation and verification steps:
1) Create production baseline
   - Repository qtqa/tqtc-coin-ci production branch HEAD is checked
     out followed by git merge origin/master
2) Test production baseline
   - Run repository auto-tests with make citest
   - Schedule integration tests with run_builds.py cli

If tests are successful, the user should amend the commit with test
results and push the change to qtqa/tqtc-coin-ci repository (requires
rights to push merges).

NOTE: If any of the tests in step 2 fail, this should be mentioned
in the commit message of the merge.

Change-Id: I5f44f1a3ee9f52771b25e8ea6669978e90277d29
Reviewed-by: Simo Fält <simo.falt@qt.io>