Fix setup.pl to work with newer versions of perl5
Setup.pl was using old version of cpanminus (1.4005) bootstrap script
which did not work with newer perl 5 versions. Newer versions of
perl5 had some problems with regex used in older version of the
script. Cpanmminus 1.7044 bootstrap script no longer has this
problem, should work also with windows. Also website had to be
changed as search.cpan.org has shutdown as of 25 of June 2018
metacpan.org is successor site to it.

Change-Id: I942d0335757ab7afe7dde915a75e9fc80367473e
Reviewed-by: Aapo Keskimolo <aapo.keskimolo@qt.io>