Script for enabling automatic extraction of replicated build artifacts
Coin will replicate wanted build artifacts and source files over to
the file server. These packages are in .tar format and need to be
extracted.

This change adds a script for enabling the incrond job for automatic
extraction of the .tar packages sent over to the file server.

Task-number: QTQAINFRA-1938
Change-Id: I06aff0ba5feea740bbf51633dca910df470055bf
Reviewed-by: Aapo Keskimolo <aapo.keskimolo@qt.io>