Fix re-running the bot when dependencies were changed manually
It may happen that a dependencies update was merged by hand, in which case the script should catch up with reality
and see that the update isn't necessary. This happens automatically when the yaml file
in the repo is identical with what the bot proposes. That however in turn requires us to make
sure that we're up-to-date, so refresh the module tip before doing so.

Change-Id: I352eba235440ab0ed38bafa9f07b07be8aed0472
Reviewed-by: Lars Knoll <lars.knoll@qt.io>