branch_qt: Fix merge and sync to update staging branches
Staging branches need to be updated to match the new target branches so that CI
will work correctly.

Now both merge and sync will also avoid pushing while there are changes staged
or integrating in the target branch. Version bumps were already taking care of
not pushing in that situation.

Change-Id: I1d3bde14cc367429721873043ef54544ef43a90d
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>