Prospective fix for shadertools update deadlock
Shadertools is a dependency of qtquick3d (per yaml file), but it's not
in qt5.git yet. So the update ends up with dependency that's never
satisfied. The shadertools can't be added to qt5.git yet because it
needs an update first. So it's a deadlock.

To fix this, allow specifying a product ref in autorun.json for fetching
the list of qt5 modules. This set to the change that adds shadertools
gives the correct list of modules.

Change-Id: I03265dd4023b17d2b2860be2170c81a34c47c5aa
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>