Say hello to cherry-pick automation
As part of the new branching model proposal discussed at the 2019
qt contributors summit, this patch introduces the basic framework
upon which the bot will be built. This patch contains basic server
components for listening to incoming requests and a notifier framework
for transiting notifications between bot subsystems. The toolbox serves
as a generic container for helper methods.

Task-number: QTQAINFRA-3382
Change-Id: I773b32e66fc505f2756ec27e82e51a0118cd3b78
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>