Add gerrit rest tools to cherrypick bot
This patch adds a suite of REST based tools which can be used for
various query types to gerrit, including setting approvers, creating
cherry picks, and so on. Also added is a basic queuing structure for
cherrypick state tracking which can be used to write updates to the
database in a first-in-first-out manner.

Task-number: QTQAINFRA-3382
Change-Id: Ibb0a5e522bfe61e5ec7d33728d1e882da8ff15f6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>