Add phase 2 functionality to cherry-pick bot: Handle relation chains
Phase 1 of the cherry-pick bot focused on the core framework
for handling a single change with one or more cherry picks.

Phase 2 (this patch) includes a great deal of additional functionality
that enables the bot to manage cherry-picks that are included
as part of a relation-chain. The bot now checks incoming requests
for related changes, determines if the change is ready to be picked,
and then performs the pick when all criteria is met (the origin branch
parent has a pick that is in a state that can be used as a
target parent for the new pick). A system of event listeners and
emitted events is created for any change which has unmet
pre-requisites. In this way, the pick order of changes in a chain
is maintained, and conflict-free changes are staged together where
possible. For changes which have unmet pre-requisites for picking,
comments are posted to the appropriate change in gerrit. The bot
will then listen for 48 hours for the pre-requisite to be met. If the
condition is met during this period, the bot will resume the process
immediately. If the window lapses, a follow-up comment will  be posted
asking the user to take a specified manual action.

This is a squashed patch that includes the following major changes:
Make Notifier a generic event dispatcher
Split single change and relation chain logic into separate managers
Add parent cherry-pick validation
Add staging eligibility check
Add event emitters for internal actions
Add gerrit actions to set the assignee of a cherry pick
Add gerrit queries to pull a revision's relation chain
Fix issues relating to Heroku compatibility
Use prepared queries on SQL statements in cherry-pick bot

Task-number: QTQAINFRA-3382
Change-Id: I136464b1fd0385407a3d30dc1e54d47a4ae0e66e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>