Add Phase 3 functionality to cherry-pick bot: Process restoration
Phase 1 of the cherry-pick bot focused on the core framework
for handling a single change with one or more cherry picks.

Phase 2 focused on enabling handing of changes which were part
of a relation chain.

Phase 3 (this patch) focuses on adding logic to restore processes
if the bot were to be restarted. This is accomplished by storing
state data in the database for each incoming change-merged request,
and using the event-driven nature of the cherry-pick bot to "re-emit"
the last-known signal an in-process item had used. Since the bot is
also capable of listening for events and updating otherwise "finished"
items, these listeners are also stored in the database and restored
upon startup.

This is a sqashed patch of the following major changes:
Overhaul cherry-pick bot logging system
Add process restoration to cherry-pick bot
Remove finished_requests table and do not move completed items
Update cherry-pick bot handling logic to search for a suitable parent
Update cherry-pick commit message parser to also handle 'dev'
Reduce boilerplating of gerrit URLs and authentication
Replace config variable handling to be more concise

Formatting has been modified from the original patch sources.

Task-number: QTQAINFRA-3382
Change-Id: I00c156ce64b80b07184fb5213420f6b37f9de7d1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>