Correct and simplify API change review script
The fallback branch when sync-filter-api-headers doesn't help (which
should seldom be used) neglected to echo files that weren't to be
filtered out.

The help message failed to mention that sub-modules are always left on
the prior release's version.

When implementation of AMEND was not needed, the check that AMEND is
set was needlessly duplicated.

Change-Id: Ibfd2161417fc035bc6659566e5fe391741c96b43
Reviewed-by: Paul Wicking <paul.wicking@qt.io>