Provide some improvements to testwheel.py
The testwheel.py script gave some problems on macOS:

macOS produced a lot of error messages which said that
libshiboken was not found. Then the script finished without
any output but those error messages.

After quite a long search, I recognized that the script was
only building on macOS without running. Running was explicitly
disabled for some unknown reason.

I added some explicit messages that the test run was ok and
also enabled running the translated script.
The warning level was set to critical for macOS to get rid of
the irritating errors that are no errors.

Further, I also allowed testing on Python 2 without PyInstaller.

Change-Id: I334071d66c0996f1abca97c98141cccbdcf2637e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>