testwheel.py: Use correct path expression instead of p.endswith()
The standard way to check how a path specification ends is

    os.path.basename(p) == some_str

The expression

    p.endswith(some_str)

is wrong, because it would also work for a longer string.

My previous solution to use a slash in the string was really bad,
the proposal of some version of

    p.endswith(os.sep + some_str)
    p.endswith(os.path.join(os.sep), some_str)

was even worse and totally ignored the well know standard
solution. Sometimes the obvious seems to be hidden ;-)

Change-Id: Ie65bf4acade16fe8c2ba03546c7806f488808acc
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>