Cherry-pick Bot: Fix listener logic to avoid race conditions
An issue was observed where a change that was part of a relation
chain would be merged at the same time as its parent, but was not
dependent on the parent. In such a situation, the resulting cherry-
pick of the child change would be conflict-free and stage
automatically. If the merge notification for the parent change is
received by the bot first, things work as expected. If the merge
notification of the non-dependent is received first however,
the child is cherry picked and staged, but is still listening
for the parent cherry-pick to be created. When it is, just miliseconds
later, the bot would try to upload a new patchset and stage again.
This fails because the non-dependent child is already in a staged
status.

The new logic works around this type of issue by introducing cancel
events to the listener logic. In the example above, the staging event
would cancel the listener for a new parent, since the change appears
to be independent from its original parent.

In addition to the cancel logic of listeners introduced, alert messages
posted to gerrit may also be delayed until a condition is met. This
prevents informational messages that may call a user to action from
being posted until the bot has confirmed that action is truly needed.

Task-number: QTQAINFRA-3382
Change-Id: I358c6c89214e0df067364393fe24b9ec6b762dbf
Reviewed-by: Paul Wicking <paul.wicking@qt.io>