Add a framework for extending cherry-pick bot with arbitrary behavior
This patch adds the necessary components for cherry-pick bot to
serve as a core of functionality which will support the easy and
rapid development of plugin bots with access to easy logging,
database tools, and simple usage of the gerrit REST API.

For each subdirectory in "plugin_bots", a JS file of the same name
of the subdirectory will be instantiated with access to Notifier
and Logger. Through Notifier, the plugin bot will gain access to:
Server, RequestProcessor, and RetryProcessor instances, along
with the ability to bind new listener paths in Notifier to receive
events relevant to the plugin bot, and custom event emission through
Server on incoming messages from gerrit.

This change also updates a few helper methods to assist in
extensibility.

Task-number: QTQAINFRA-3779
Change-Id: I848ad158b99478e264f595dc6b60093b57c9311c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>