Cherry-pick bot: Retry last action on gerrit 409 LockFailureException
Gerrit sometimes fails to release resource locks quickly enough for
external automation. If a POST/PUT request to gerrit results in
a 409 LockFailureException, retry the action with a delay until
it succeeds or throws a different error.

Additionally, force a 5 second delay before sending the request to
stage a change, since this delay has been observed in testing to
reduce the number of 409 Lock failures.

Fixes: QTQAINFRA-3780
Change-Id: I39cabbbec3b4ddb463ce97b27dab9491b25f47e2
Reviewed-by: Paul Wicking <paul.wicking@qt.io>