Fix handling of new files in resetboring.py's disclaimed list
In api-review-gen, simply piping the disclaimed files into
  xargs -r git checkout $QUIET "$PRIOR" --
ran into trouble with new files, which got error message
  filespec '...' did not match any file(s) known to git
and weren't removed; this also caused other files in the same
invocation of git checkout to be left unreverted.

So loop over the files, doing the git checkout on each and falling
back to || rm -f for any on which that fails; and log the rm -f when
running it, so that the filespec error is followed by evidence that it
was sorted out.

Task-number: QTBUG-86400
Change-Id: I9d0f9f5766aa4a43f85692ab0723493bf9aeebb5
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>