Limit the chattiness of the checkout that removes old headers
The checkout of old headers on the new branch, with git rm when they
don't exist there, was passing -q for the checkout but using $QUIET in
the git rm.  Replace both with a partner to $QUIET, $UNQUIET, that
defaults to -q but is cleared by -v so that this (generally dull) list
of removals isn't reported by default, but can be enabled.

Task-number: QTBUG-86400
Change-Id: I5e1c48a1b0aac9e2364279d6bfed66372c683b21
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>