Submodule update bot: Allow manual selection of qtbase sha
It's possible that when starting a new round, it may be desirable
to select a specific qtbase integration sha, instead of HEAD.
This typically occurs when a submodule update round has trouble
moving forward because too many breaking changes have been applied
to qtbase.
This change allows manual section of a valid sha to serve as the
starting point for a new round.

New environment variables:
Set CUSTOM_QTBASE to select a custom qtbase without user input.
Set AUTORUN to choose qtbase/head without user input.
If neither is set, user input is required to input a sha or
accept the current head.

Change-Id: Ice633151e06797749028c9e06073563b137548d8
Reviewed-by: Toni Saario <toni.saario@qt.io>