Cherry-pick Bot: Fix retry action on parent merged race condition
When a parent and child are merged at the same time, and both must
create cherry-picks to other branches, a race condition may
occur where the cherry picks for the child will be created first.
In this instance, we should wait a few seconds and check to see
if the parent has been created.

This change fixes an issue where the isRetry flag does not get
passed to the function that checks for the parent, and the code
path dead-ends.

Change-Id: Ia60cf5920c60a28101c03973bc2b497536e33eff
Reviewed-by: Paul Wicking <paul.wicking@qt.io>