Create a voting system for use in the Qt Governance Model
A simple voting system is needed to provide anonymous, but
authenticated voting which limits the voting to users
in the Approvers and Maintainers groups within The Qt project.

This plugin for the cherry-pick-bot framework provides such a
voting system. As a plugin to the framework, it can be enabled
and configured on-demand, and then disabled when no longer needed.
Votes are stored in a separate table of the cherry-pick-bot's
primary database. In order to satisfy GDPR compliance, this database
is intended to be be wiped shortly after the voting round completes.
Logging/tracing tracks only who votes and when, but does not
record votes.

If users are not comfortable with using their gerrit API
password for authentication, they are directied to email
gerrit-admin@qt-project.org to request a personal voting token
which will be associated with their gerrit user in the voting
database.

Change-Id: Ica7792aede21d3eaa51828414afd4ed6805baf40
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>