Revert "oss-fuzz: Don't set platform explicitly"
It fixed building Qt with afl++ but makes qmake unusable due to
QTBUG-90931.

Reverting the change will allow building with clang because there, the
compilers set by "-platform" and CC/CXX are the same. This will give me
a little time to either work around that issue or port the tests to
cmake.

This reverts commit c50683e9d8b96112366af8dab91403cdc877a221.

Task-number: QTBUG-90931
Change-Id: I1cef6f3002d940511e1844d76787978c2d3149b9
Reviewed-by: Kai Koehne <kai.koehne@qt.io>