Fix event listener restoration functionality
Due to an extra parameter being passed to the
addToListenerCacheUpdateQueue function during updateDBListenerCache,
the cache update queue was being flushed. While this still
allowed listeners to be set up initially, they were not being
persisted to the database. As such, only the first listener in queue
would be set up fresh upon restarting the bot, with default timeouts,
and erroneously re-posting messages to gerrit with each bot restart.
This had a side effect of listening for events much longer than
intended, as heroku restarts the bot at least daily.

Also included in this patch is a fix to stop processing a changes
when its parent has not yet merged. In such a case, listeners
for the parent's events should already be set up. If the bot
restarts at this point, work will be duplicated as the change will
both be reprocessed by the bot as usual, and by the listener if the
expected parent event occurs. All work should be left to the event
listener at this point.

Change-Id: Ib210e883d7d4a619cf0adb8b3d62eab1dea6af24
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>