Add a script to help collect statistics on restaging
This script processes merged changes on codereview and scrapes
comments for information about how many times a change failed CI
before passing, or if it was rejected due to a conflict in the
staging branch.

Data is collected by repo and branch, then written to an instance
of influxdb for further processing.

By default, the script generates results for essential and addon
modules of qt5, as well as submodule updates. Additional statistics
are collated for these super-groups to reduce the work required
with the database to get an overview of the larger data set.

Task-number: QTQAINFRA-4411
Change-Id: I4c857d82c915fade3614ef6b4f2a74f4986daf79
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>