Add API Change Review wrapper script
This is a wrapper script that parses .gitmodules of a qt5 checkout and
figures out which version it should be compared to. It then generates a
change review for each module and asks if it should be uploaded. It acts
as a replacement for all-api-modules which also handles the steps of
checking out repos and the otherwise manual git steps.

Change-Id: If2035fe70159d33fad7f5f6375f5ef1ca5e6a8ea
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>