Blacklist QBindingStatus and QQmlPrivate::AOTCompiledFunction
It's a bit unfortunate that we need this. QBindingStatus is not used in a way
that would require its size to never change, it has had data members added to
it. The BIC test doesn't have a way to blacklist it in a way that would just
keep its member offsets stable. We should eventually rewrite all of this
to use libabigail, and add support for being able to say that a class is bic-stable
as long as its existing data members don't move.

Pick-to: master
Change-Id: I97b13236249a76dc6ab24037d4d8d29611356f40
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>