Rework QtBranching.bump_qtbase_datastream()
The enum in qdatastream.h now uses QT_VERSION_CHECK() instead of overt
hex equal to it (but support recognising the old form, just in case -
it'll duly convert to use the macro, which it should). Added a
doc-string to the function, explaining what it's doing.

The lines in the files sometimes get reordered; make the substitutions
more robust against that by matching single lines. The matched lines
should match the prior version, so actually check for that. The new
version's enum member should be set to *the same one* the prior
version is set to, and documented as such, rather than to the prior
version itself. (This makes no difference to the compiler, but is
clearer for folk reading the code and docs.)

Use raw strings to save repeated backslashes and simplify some
regexes.

Change-Id: I0cf504a8d8dc24c9ce8e47a90437bed6efbfe1cc
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>