Cache incoming events and check the cache after setting up listeners
Race conditions may occur when multiple changes arrive at the bot
simultaneously where a cherry-pick begins listening for an event.
Before the listener can be set up and begin listening, the target
event arrives, causing the cherry-pick to miss the event.

This patch adds an event cache such that incoming events from gerrit
are stored for 30 seconds before being cleared. When an event listener
is set up, it now checks for the event to listen for in the cache and
rebroadcasts it, triggering the configured listener. The event in the
cache is then allowed to expire. Duplicate incoming events will
reset the expiry time of a given event in the cache.

Fixes: QTQAINFRA-4409
Change-Id: Ie7ac91974cc7522cd65b0497b309820e1de16da3
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>