Add an option to co-stage version bumps with the submodule update bot
Use the option "--automatic-stage" to enable self-approving of created
version bump stages, along with adding
 qt_submodule_update_bot@qt-project.org as a reviewer.
This will enable the normal submodule update rounds to stage version
bumps with an update round. Co-staging version bumps with submodule
updates ensures a clean and consistent set of Qt modules when the
version changes, so even the first revision of Qt X.x.x is buildable
and passes tests.

Change-Id: If781c198aa47100a9f94bb656fc6547bcca6c078
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>