Dedupe reporting of failed tests
In the CI, we rerun tests if they fail. The reporting script would have
reported each rerun, which creates quite a bit of line noise. Use a set
to deduplicate them. The reruns' outputs differ by a number indicating
the run count, so use a regex to remove that one.

Pick-to: master
Change-Id: I65c11847375434dfd1470e4a9456396029567d0e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>