Update cherry-pick bot to add users to the Attention Set
With the gerrit upgrade to 3.3.7, the assignee field is no longer
visible in the UI and will be deprecated in the future. This patch
updates the bot and its plugins to use the Attention Set instead.
Users can be added to the attention set with a reason note. Changes
for which a user is tagged on will appear in that user's Gerrit home
page as highlighted entries, better bringing attention to them.

Task-number: QTQAINFRA-4730
Change-Id: If77cc45cda45f3d90e2bf14883d60a74b1046225
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>