Add a plugin to monitor changes created by cherry-pick-bot
This plugin monitors for new patchsets and integration failures
among changes owned by the cherry-pick-bot.
When a user who is not the original author uploads a new patchset,
this plugin adds the original author or approving reviewer to the
attention set of the bot-owned cherry-pick.
When a bot-owned cherry-pick fails to integrate, the author
of the most recent patchset or of the original change is
added to the attention set.

This updates the current behavior where the cherry-pick-bot
in most cases forgets about a change when it is successfully
staged or has a merge conflict. This behavior has sometimes
resulted in cherry-picks getting lost, as the cherry-pick-bot
becomes the owner of new cherry-pick changes it creates and
the original author is never notified of problems.

Task-number: QTQAINFRA-4611
Change-Id: I3ab61e38aadb31ff5ce0ab7bf0d13b601340890e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>