tst_license.pl: Exclude 3rdparty directories for CI too
tst_license.pl has two ways of compiling file lists:
- git ls-files for git repos
- find otherwise

The CI works on source archives, not git repos.

Commit 58abc488767a0a0d7e79a67a393ab89ada635b00 tried to exclude all
3rdparty files but did so only for the git repo case.

tst_license.pl already checked for 3rdparty files and classified them as
"optional".  Those files are checked for valid licenses if they contain
something that looks like a license.  This is now problematic for
porting to SPDX, because there are 3rdparty files with an SPDX license
identifier but without a copyright tag.  This combination is illegal in
terms of the REUSE specification, and we must exclude those files.

Now, we classify 3rdparty files as "excluded" instead of "optional",
regardless of working on a git repo or source archive.

Pick-to: dev
Change-Id: Ie6f4f3e2ede39de2f3872ae588fe960323d15122
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>