Set topic on changes cherry-picked from tqtc/ shadow repos
When a pick is requested in an LTS branch which resides in
a tqtc/ shadow repository, the pick is performed by using the
bare sha of the originally merged change in the shadow repo.
Changes are not replicated to shadow repos, only git objects,
so this limitation is expected. Since git objects do not retain
topics as set in gerrit changes, the cherry-pick bot must retrieve
the topic from the original change and set it when performing
the cherry-pick action in tqtc-private repos.

Since this is the same behavior as during a normal cherry-pick via
a change anyway, this fix always sets the topic regardless of
cherry-pick source with no ill effect.

Fixes: QTQAINFRA-4987
Change-Id: I8424ba8c40291b7a003d4e326a4d45da02bb5881
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>